numeros float printados com 2 casas decimais

+++ src/com/company/database/Database.java
@@ -92,7 +92,7 @@ public class Database {
             totalLateDays += lateDays;
             totalLocations++;
         }
-        average = totalLateDays / totalLocations;
+        average = (float)totalLateDays / totalLocations;
         return average;
     }
